requirements.txt
tkinter
Pillow>=9.0.0
configparser 